Migrate examples to use esbuild instead of parcel
