+++ examples/webcam/index.html
@@ -22,7 +22,7 @@
       An example of using a webcam for upscaling with
       <a target="_blank" href="https://github.com/thekevinscott/UpscalerJS">UpscalerJS</a>.
     </p>
-    <video autoplay="true" id="video" width="400" height="300"></video>
+    <video autoplay="true" id="video" width="200" height="150"></video>
     <p>Click the upscale button to upscale a frame of video.</p>
     <button id="button" disabled>Upscale</button>
     <p id="info"></p>
