+++ examples/webworker/src/animate.js
