+++ examples/webworker/src/flower.png
