+++ examples/webworker/src/index.html
@@ -46,7 +46,7 @@
     <tbody>
       <tr>
         <td>
-          <img src="/flower-256.png" id="flower" />
+          <img src="/flower.png" id="flower" />
         </td>
         <td>
           <div id="target"></div>
