+++ examples/webworker/src/worker.js
