+++ packages/upscalerjs/src/index.ts
@@ -1,3 +1,2 @@
 export { default, } from './upscaler';
 export { AbortError, getRowsAndColumns, getTensorDimensions, } from './upscale';
-alert('1');
